backend/requirements.txt
faiss-cpu
sentence-transformers
requests
numpy
pickle-mixin

flask
groq
python-dotenv
